+++ space-d/src/main/java/com/dnd/spaced/core/admin/application/event/listener/AdminWordEventListener.java
@@ -1,7 +1,7 @@
 package com.dnd.spaced.core.admin.application.event.listener;
 
 import com.dnd.spaced.core.admin.application.event.dto.DeletedWordEvent;
-import com.dnd.spaced.core.word.application.DeletedWordIdRepository;
+import com.dnd.spaced.core.word.application.repository.DeletedWordIdRepository;
 import com.dnd.spaced.core.word.domain.repository.PronunciationRepository;
 import com.dnd.spaced.core.word.domain.repository.WordExampleRepository;
 import java.time.Clock;
